+++ packages/plugin-invalidation/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @spoosh/plugin-invalidation
 
+## 0.7.1
+
+- Fix params not resolving correctly
+
 ## 0.7.0
 
 - Move `invalidate` option to new write trigger options